requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-model-utils==4.1.1
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
slack-sdk==3.4.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
